stop showing stage from in new project, it is confusing
