+++ app/views/projects/new.html.erb
@@ -1,4 +1,4 @@
-<h1>Setup a new project</h1>
+<h1>New project</h1>
 
 <section>
   <%= render "form" %>
