+++ test/integration/cleanliness_test.rb
@@ -93,7 +93,6 @@ describe "cleanliness" do
   it "tests all files" do
     untested = [
       "app/controllers/concerns/current_project.rb",
-      "app/controllers/concerns/stage_permitted_params.rb",
       "app/mailers/application_mailer.rb",
       "app/models/changeset/code_push.rb",
       "app/models/changeset/jira_issue.rb",
